(Re)Organize flex message response. Keep sample-spring-boot-echo simple.
Move flex message response test from sample-spring-boot-echo to sample-spring-boot-kitchensink.
KEEP sample-spring-boot-echo SIMPLE.
