+++ sample-spring-boot-kitchensink/src/main/java/com/example/bot/spring/ExampleFlexMessageSupplier.java
@@ -14,7 +14,7 @@
  * under the License.
  */
 
-package com.example.bot.spring.echo;
+package com.example.bot.spring;
 
 import static java.util.Arrays.asList;
