+++ sample-spring-boot-kitchensink/src/main/java/com/example/bot/spring/KitchenSinkController.java
@@ -418,6 +418,9 @@ public class KitchenSinkController {
                         )
                 ));
                 break;
+            case "flex":
+                this.reply(replyToken, new ExampleFlexMessageSupplier().get());
+                break;
             default:
                 log.info("Returns echo message {}: {}", replyToken, text);
                 this.replyText(
